rename services/platform.service.js - consolidate import statements in task middleware - pin ps-tree version in package.json
