+++ package.json
@@ -39,7 +39,7 @@
     "electron-store": "2.0.0",
     "fix-path": "2.1.0",
     "gatsby-cli": "1.1.58",
-    "ps-tree": "^1.1.0",
+    "ps-tree": "1.1.0",
     "yarn": "1.9.2"
   },
   "devDependencies": {
