+++ src/services/dependencies.service.js
@@ -1,5 +1,5 @@
 // @flow
-import { PACKAGE_MANAGER_CMD } from './platform.services';
+import { PACKAGE_MANAGER_CMD } from './platform.service';
 import * as childProcess from 'child_process';
 
 const spawnProcess = (cmd: string, cmdArgs: string[], projectPath: string) =>
